feat: allow skipping jobs already assigned jobs (#236)
* feat: allow skipping jobs already assigned jobs

This adds a `-u/--unassigned` flag to `ght assign` to allow for skipping
reassignment of jobs that are already assigned. This is intended to
avoid spamming interviewers when add a new candidate to the queue.

* fix: set user agent in development mode

When running in development mode, you may encounter a 429 too many requests
on page load. Setting a user agent works around this.

* chore: clean up user agent comment for clarity