Accept the beforeunload handler and skip excessive reload of the original job post (#273)
* Skip excessive reload of the original job post

The expected URL is already opened when replicating to the second city
and on. Also, by skipping the excessive reloads, it avoids the
beforeunload handler of the site too since we are not leaving the page
quickly and frequently.

Closes: #268

* Accept the beforeunload handler when leaving the edit page

"/jobapps/${jobPost.id}/edit" may have a beforeunload handler so accept
the "Changes you made may not be saved" dialog since we are not editing
anything but replicating the existing data only.

Closes: #268

* move page dialog registration to make it called once only.

* version bump

---------

Co-authored-by: Marius Kavaliauskas <mariuskava@canonical.com>